+++ changes.md
@@ -25,12 +25,27 @@ This release ...
   microenvironment.diffusion_decay_solver = diffusion_decay_solver__constant_coefficients_LOD_1D
   
   Use this right after setup_microenvironment() in your main.cpp file. Future versions will include an XML option to use 1D. Most users will never need this. 
+  
++ Added a standardized chemotaxis function to the standard models: 
+
+     void chemotaxis_direction( Cell* pCell , Phenotype& phenotype , double dt ); 
+
+  This sets:
+  
+     phenotype.motility.motility_bias_direction = direction * grad( index ), where 
+    
+     direction = phenotype.motility.chemotaxis_direction 
+        (1 to go up gradient, -1 to go down gradient)
+     index = phenotype.motility.chemotaxis_index 
+        (the index of one of hte diffusing substrates)
 
 + Cell_Definitions in XML. This is in response to SourceForge ticket 5. 
    +
    +
 
 ### Minor new features and changes: 
+
++ added int chemotaxis_index and chemotaxis_direction to the Motility class to assist with a new standard chemotaxis function. 
  
 + scale_all_secretion_by_factor also scales net_export_rates.